Bulk rename Eigen -> QRKit
