+++ src/QRKit/BlockedThinSparseQR.h
@@ -8,15 +8,15 @@
 // Public License v. 2.0. If a copy of the MPL was not distributed
 // with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#ifndef EIGEN_BLOCKED_THIN_SPARSE_QR_H
-#define EIGEN_BLOCKED_THIN_SPARSE_QR_H
+#ifndef QRKIT_BLOCKED_THIN_SPARSE_QR_H
+#define QRKIT_BLOCKED_THIN_SPARSE_QR_H
 
 #include <ctime>
 #include <typeinfo>
 #include <shared_mutex>
 #include "BlockedThinQRBase.h"
 
-namespace Eigen {
+namespace QRKit {
 
   template<typename MatrixType, int SuggestedBlockCols, bool MultiThreading> class BlockedThinSparseQR;
 
@@ -287,6 +287,6 @@ namespace Eigen {
     std::vector<Index> m_zeroColPermIdxs;
   };
 
-} // end namespace Eigen
+} // end namespace QRKit
 
 #endif
