+++ test/test-qrkit.cpp
@@ -24,9 +24,10 @@
 
 #include <Eigen/Eigen>
 #include <Eigen/SparseCore>
-#include <Eigen/SparseQRExtra>
+#include <QRKit/QRKit>
 
 using namespace Eigen;
+using namespace QRKit;
 
 typedef double Scalar;
 
@@ -58,7 +59,7 @@ void generate_overlapping_block_diagonal_matrix(const Eigen::Index numParams, co
   std::uniform_real_distribution<double> dist(0.5, 5.0);
 
   int stride = 7;
-  Eigen::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams);
+  QRKit::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams);
   for (int i = 0; i < numParams; i++) {
     for (int j = i * 2; j < (i * 2) + 2 && j < numParams; j++) {
       jvals.add(i * stride, j, dist(gen));
@@ -96,7 +97,7 @@ void generate_block_diagonal_matrix(const Eigen::Index numParams, const Eigen::I
   std::uniform_real_distribution<double> dist(0.5, 5.0);
 
   int stride = 7;
-  Eigen::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams);
+  QRKit::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams);
   for (int i = 0; i < numParams; i++) {
     for (int j = i * 2; j < (i * 2) + 2 && j < numParams; j++) {
       jvals.add(i * stride, j, dist(gen));
@@ -130,7 +131,7 @@ void generate_block_angular_matrix(const Eigen::Index numParams, const Eigen::In
   std::uniform_real_distribution<double> dist(0.5, 5.0);
 
   int stride = 7;
-  Eigen::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams + numResiduals * numAngularParams);
+  QRKit::TripletArray<Scalar, typename JacobianType::Index> jvals(stride * numParams + numResiduals * numAngularParams);
   for (int i = 0; i < numParams; i++) {
     for (int j = i * 2; j < (i * 2) + 2 && j < numParams; j++) {
       jvals.add(i * stride, j, dist(gen));
